:tractor: Move models and types into their own package
